NLPtoSQL/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pymysql
pymssql
python-dotenv
pydantic
langchain
langchain-community
langchain-groq
cryptography
